Remove FABS3.1 and add FABSREQ11
