+++ dataactvalidator/config/sqlrules/fabsreq11_detached_award_financial_assistance.sql
@@ -1,13 +1,10 @@
--- Action type is required for non-aggregate and PII-redacted non-aggregate records (i.e., when RecordType = 2 or 3)
+-- ActionType is required for all submissions except delete records, but was not provided in this row.
 SELECT
     row_number,
     action_type,
-    record_type,
+    correction_delete_indicatr,
     afa_generated_unique AS "uniqueid_AssistanceTransactionUniqueKey"
 FROM detached_award_financial_assistance
 WHERE submission_id = {0}
-    AND record_type IN (2, 3)
-    AND (action_type = ''
-        OR action_type IS NULL
-    )
+    AND COALESCE(action_type, '') = ''
     AND UPPER(COALESCE(correction_delete_indicatr, '')) <> 'D';
