+++ dataactvalidator/config/sqlrules/sqlRules.csv
@@ -86,7 +86,6 @@ FABS2.1,"The combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Nu
 FABS2.2.1,"The combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Number, and AwardingSubTierAgencyCode must be unique when compared to currently published data--unless the record is a correction or deletion (i.e., if CorrectionDeleteIndicator = C or D). In this particular case, the combination of these five fields in this transaction has already been published in USAspending, making this second attempt a duplicate.",n,fabs,fatal,,fabs2_detached_award_financial_assistance_2_1,"The combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Number, and AwardingSubTierAgencyCode must be unique when compared to currently published data.","existence"
 FABS2.2.2,"The unique combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Number, and AwardingSubTierAgencyCode must exist as a currently published record when the record is a correction (i.e., if CorrectionDeleteIndicator = C).",n,fabs,fatal,,fabs2_detached_award_financial_assistance_2_2,"The combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Number, and AwardingSubTierAgencyCode must be unique when compared to currently published data.","existence"
 FABS2.2.3,"The unique combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Number, and AwardingSubTierAgencyCode must exist as a currently published record when the record is a deletion (i.e., if CorrectionDeleteIndicator = D).",n,fabs,fatal,,fabs2_detached_award_financial_assistance_2_3,"The combination of FAIN, AwardModificationAmendmentNumber, URI, CFDA_Number, and AwardingSubTierAgencyCode must be unique when compared to currently published data.","existence"
-FABS3.1,"ActionType is required for non-aggregate and PII-redacted non-aggregate records (RecordType = 2 or 3).",n,fabs,fatal,,fabs3_detached_award_financial_assistance_1,"A, B, C, or D","completeness"
 FABS3.2,"ActionType must contain one of the following values: ""A"", ""B"", ""C"", ""D"", or blank.",n,fabs,fatal,,fabs3_detached_award_financial_assistance_2,"A, B, C, D, or (blank)","completeness"
 FABS4.1,"ActionDate must follow YYYYMMDD format, where YYYY is the year, MM the month and DD the day.",n,fabs,fatal,,fabs4_detached_award_financial_assistance_1,"ActionDate must follow YYYYMMDD format","accuracy"
 FABS4.2,"Value of ActionDate must be between 19991001 and 20991231 (i.e., a date between 10/01/1999 and 12/31/2099).",n,fabs,fatal,,fabs4_detached_award_financial_assistance_2,"Between 10/1/1999 and 12/31/2099","accuracy"
@@ -184,3 +183,4 @@ FABSREQ6,"BusinessTypes is required for all submissions except delete records, b
 FABSREQ7,"AssistanceType is required for all submissions except delete records, but was not provided in this row.",n,fabs,fatal,,fabsreq7_detached_award_financial_assistance,"AssistanceType must not be blank","completeness"
 FABSREQ8,"ActionDate is required for all submissions except delete records, but was not provided in this row.",n,fabs,fatal,,fabsreq8_detached_award_financial_assistance,"ActionDate must not be blank","completeness"
 FABSREQ9,"AwardeeOrRecipientLegalEntityName is required for all submissions except delete records, but was not provided in this row.",n,fabs,fatal,,fabsreq9_detached_award_financial_assistance,"AwardeeOrRecipientLegalEntityName must not be blank","completeness"
+FABSREQ11,"ActionType is required for all submissions except delete records, but was not provided in this row.",n,fabs,fatal,,fabsreq11_detached_award_financial_assistance,"ActionType must not be blank","completeness"
