+++ tests/unit/dataactvalidator/test_fabsreq11_detached_award_financial_assistance.py
@@ -1,0 +1,35 @@
+from tests.unit.dataactcore.factories.staging import DetachedAwardFinancialAssistanceFactory
+from tests.unit.dataactvalidator.utils import number_of_errors, query_columns
+
+_FILE = 'fabsreq11_detached_award_financial_assistance'
+
+
+def test_column_headers(database):
+    expected_subset = {'row_number', 'action_type', 'correction_delete_indicatr',
+                       'uniqueid_AssistanceTransactionUniqueKey'}
+    actual = set(query_columns(_FILE, database))
+    assert expected_subset == actual
+
+
+def test_success(database):
+    """ Test ActionType is required for all submissions except delete records. """
+
+    det_award = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr='C', action_type='c')
+    det_award_2 = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr='', action_type='A')
+    # Test ignoring for D records
+    det_award_3 = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr='d', action_type=None)
+    det_award_4 = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr='D', action_type='')
+    det_award_5 = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr='d', action_type='Name')
+
+    errors = number_of_errors(_FILE, database, models=[det_award, det_award_2, det_award_3, det_award_4, det_award_5])
+    assert errors == 0
+
+
+def test_failure(database):
+    """ Test fail ActionType is required for all submissions except delete records. """
+
+    det_award = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr='c', action_type=None)
+    det_award_2 = DetachedAwardFinancialAssistanceFactory(correction_delete_indicatr=None, action_type='')
+
+    errors = number_of_errors(_FILE, database, models=[det_award, det_award_2])
+    assert errors == 2
